+++ src/feature/auth/ui/index.ts
@@ -1,1 +1,1 @@
-export * from "./PasswordChangeModal";
+export * from "./LoginForm";
