+++ src/feature/passwordChange/index.ts
@@ -1,0 +1,3 @@
+export * from "./ui/index";
+export * from "./model/index";
+export * from "./service/index";
